fix(wallet): gate dApp RPC on granted permission
The handler ended in a default: branch forwarding any unrecognised method
to the node for any origin, connected or not. Dispatch now mirrors
status-go's CallRPC: local registry, remote allowlist, then -32601.

Method bodies move to lib/rpc/ and rpc-handler.ts keeps a handler map, so
the gated set cannot drift from the implemented one.
