+++ apps/wallet/src/lib/rpc/send-transaction.ts
@@ -1,0 +1,8 @@
+import { ProviderRpcError } from '@status-im/ethereum-provider'
+
+export async function eth_sendTransaction(): Promise<never> {
+  throw new ProviderRpcError({
+    code: 4200,
+    message: 'Not yet supported via dApp connection',
+  })
+}
